Fix build moving submission-accordion in components
