+++ frontend/src/components/submission-accordion/index.tsx
