+++ frontend/src/components/submission-accordion/vote-selector.tsx
